Fix bug caused by reusing of Node pointers
Before this fix, when going back to the input page from the output page, the Wrangler would reuse the same nodes as initially set when going from the load page to the input page. This would cause the new outputted values to be rewritten with respect to the already wrangled values, which is of course wrong. That's been fixed now by just resetting the nodes when going back to the input page.